+++ pkg/registry/storage/volumesnapshotdata/storage/storage_test.go
@@ -20,6 +20,7 @@ import (
 	"fmt"
 	"testing"
 
+	storageapiv1alpha1 "k8s.io/api/storage/v1alpha1"
 	apiequality "k8s.io/apimachinery/pkg/api/equality"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/fields"
@@ -31,6 +32,7 @@ import (
 	genericregistrytest "k8s.io/apiserver/pkg/registry/generic/testing"
 	"k8s.io/apiserver/pkg/registry/rest"
 	etcdtesting "k8s.io/apiserver/pkg/storage/etcd/testing"
+	"k8s.io/kubernetes/pkg/api/testapi"
 	api "k8s.io/kubernetes/pkg/apis/core"
 	storageapi "k8s.io/kubernetes/pkg/apis/storage"
 	"k8s.io/kubernetes/pkg/registry/registrytest"
@@ -63,6 +65,11 @@ func validNewVolumeSnapshotData(name string) *storageapi.VolumeSnapshotData {
 }
 
 func TestCreate(t *testing.T) {
+	if *testapi.Storage.GroupVersion() != storageapiv1alpha1.SchemeGroupVersion {
+		// skip the test for all versions exception v1alpha1
+		return
+	}
+
 	storage, _, server := newStorage(t)
 	defer server.Terminate(t)
 	defer storage.Store.DestroyFunc()
@@ -80,6 +87,11 @@ func TestCreate(t *testing.T) {
 }
 
 func TestUpdate(t *testing.T) {
+	if *testapi.Storage.GroupVersion() != storageapiv1alpha1.SchemeGroupVersion {
+		// skip the test for all versions exception v1alpha1
+		return
+	}
+
 	storage, _, server := newStorage(t)
 	defer server.Terminate(t)
 	defer storage.Store.DestroyFunc()
@@ -100,6 +112,11 @@ func TestUpdate(t *testing.T) {
 }
 
 func TestDelete(t *testing.T) {
+	if *testapi.Storage.GroupVersion() != storageapiv1alpha1.SchemeGroupVersion {
+		// skip the test for all versions exception v1alpha1
+		return
+	}
+
 	storage, _, server := newStorage(t)
 	defer server.Terminate(t)
 	defer storage.Store.DestroyFunc()
@@ -108,6 +125,11 @@ func TestDelete(t *testing.T) {
 }
 
 func TestGet(t *testing.T) {
+	if *testapi.Storage.GroupVersion() != storageapiv1alpha1.SchemeGroupVersion {
+		// skip the test for all versions exception v1alpha1
+		return
+	}
+
 	storage, _, server := newStorage(t)
 	defer server.Terminate(t)
 	defer storage.Store.DestroyFunc()
@@ -116,6 +138,11 @@ func TestGet(t *testing.T) {
 }
 
 func TestList(t *testing.T) {
+	if *testapi.Storage.GroupVersion() != storageapiv1alpha1.SchemeGroupVersion {
+		// skip the test for all versions exception v1alpha1
+		return
+	}
+
 	storage, _, server := newStorage(t)
 	defer server.Terminate(t)
 	defer storage.Store.DestroyFunc()
@@ -124,6 +151,11 @@ func TestList(t *testing.T) {
 }
 
 func TestWatch(t *testing.T) {
+	if *testapi.Storage.GroupVersion() != storageapiv1alpha1.SchemeGroupVersion {
+		// skip the test for all versions exception v1alpha1
+		return
+	}
+
 	storage, _, server := newStorage(t)
 	defer server.Terminate(t)
 	defer storage.Store.DestroyFunc()
@@ -149,6 +181,11 @@ func TestWatch(t *testing.T) {
 }
 
 func TestUpdateStatus(t *testing.T) {
+	if *testapi.Storage.GroupVersion() != storageapiv1alpha1.SchemeGroupVersion {
+		// skip the test for all versions exception v1alpha1
+		return
+	}
+
 	storage, statusStorage, server := newStorage(t)
 	defer server.Terminate(t)
 	defer storage.Store.DestroyFunc()
